Add files via upload
"Subsidence_final" is the code for LiDAR files editing, combining shapefiles, joining fields and calculating elevation differences of LiDAR points to determine subsidence or uplift within specifies areas.
"Near Tool" is the calculating nearest points of LiDAR data in comparing 2 separate datasets of elevation values collected at different time periods. Using the Near tool, points from the second shapefile were merged with points from the first shapefile based on a specified distance. This process included adding the FID (identification number) of the merged point, as well as the distance and coordinates of each point from the second shapefile to the attribute tables of the first shapefile.
"file_name" is to rename files within one dataset due to similarity of names between the two datasets. This enabled calculation of "near points".
"laz2las_4_pointcount" is the convert laz files to las files and count the points captured with the dataset.
"xyz2shp" is for converting xyz to shapefiles.